pyproject.toml
[project]
name = "mikrotik-upgrade"
version = "0.1.0"
description = "A minmal tool for mass upgrade Mikrotik routers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "paramiko>=4.0.0",
    "pyyaml>=6.0.3",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
[project.scripts]
mikrotik-upgrade = "mikrotik_upgrade.upgrade:main"
